Initial upload
The initial script was written to be procedural and verbose in nature so I could re-familiarize myself with Powershell. This new branch contains the main class for the object oriented version of this script which should save hundreds of lines of code once I write the logic. This will be re written in Java, Go and C# later on for fun. 

This is the class that will be used in the object oriented version of the organize files script which I may whip up tomorrow.